Fix compile errors in c++17

+++ core/3d/CCObjLoader.cpp
@@ -859,7 +859,7 @@ std::string LoadObj(std::vector<shape_t>& shapes,
             char namebuf[TINYOBJ_SSCANF_BUFFER_SIZE];
             token += 7;
 #ifdef _MSC_VER
-            sscanf_s(token, "%s", namebuf, _countof(namebuf));
+            sscanf_s(token, "%s", namebuf, (unsigned int)_countof(namebuf));
 #else
             sscanf(token, "%s", namebuf);
 #endif
